+++ backend/de.metas.banking/de.metas.banking.base/src/main/java/de/metas/banking/payment/modelvalidator/C_Payment.java
@@ -30,6 +30,8 @@ import de.metas.banking.service.IBankStatementBL;
 import de.metas.banking.service.ICashStatementBL;
 import de.metas.payment.PaymentId;
 import de.metas.payment.api.IPaymentBL;
+import de.metas.payment.api.PaymentReconcileReference;
+import de.metas.payment.api.PaymentReconcileRequest;
 import lombok.NonNull;
 import org.adempiere.ad.modelvalidator.IModelValidationEngine;
 import org.adempiere.ad.modelvalidator.annotations.DocValidate;
@@ -41,6 +43,8 @@ import org.adempiere.model.CopyRecordFactory;
 import org.adempiere.service.ISysConfigBL;
 import org.compiere.model.I_C_Payment;
 import org.compiere.model.ModelValidator;
+
+import java.util.Collection;
 
 /**
  * @author cg
@@ -113,10 +117,14 @@ public class C_Payment
 		final PaymentId paymentId = PaymentId.ofRepoId(payment.getC_Payment_ID());
 		if (!bankStatementBL.isPaymentOnBankStatement(paymentId))
 		{
-			paymentBL.markReconciledAndSave(payment);
-
 			final PaymentId reversalId = PaymentId.ofRepoId(payment.getReversal_ID());
-			paymentBL.markReconciled(ImmutableList.of(reversalId));
+
+			final ImmutableList<PaymentReconcileRequest> requests = ImmutableList.of(
+					PaymentReconcileRequest.of(paymentId, PaymentReconcileReference.reversal(reversalId)),
+					PaymentReconcileRequest.of(reversalId, PaymentReconcileReference.reversal(paymentId)));
+
+			final Collection<I_C_Payment> preloadedPayments = ImmutableList.of(payment);
+			paymentBL.markReconciled(requests, preloadedPayments);
 		}
 	}
