+++ backend/de.metas.business/src/main/java-legacy/org/compiere/model/MPayment.java
@@ -49,6 +49,7 @@ import de.metas.payment.PaymentTrxType;
 import de.metas.payment.TenderType;
 import de.metas.payment.api.IPaymentBL;
 import de.metas.payment.api.IPaymentDAO;
+import de.metas.payment.api.impl.PaymentBL;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import de.metas.util.StringUtils;
@@ -162,7 +163,6 @@ public final class MPayment extends X_C_Payment
 			//
 			setIsReceipt(true);
 			setIsApproved(false);
-			setIsReconciled(false);
 			setIsAllocated(false);
 			setIsOnline(false);
 			setIsSelfService(false);
@@ -182,6 +182,8 @@ public final class MPayment extends X_C_Payment
 			setDateTrx(new Timestamp(System.currentTimeMillis()));
 			setDateAcct(getDateTrx());
 			setTenderType(TenderType.Check.getCode());
+
+			PaymentBL.markNotReconciledNoSave(this);
 		}
 	}   // MPayment
